refactor: reorganize header files

+++ srcs/int/ft_dup_int_arr.c
@@ -6,16 +6,17 @@
 /*   By: dcastor <dcastor@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/18 23:08:41 by dcastor           #+#    #+#             */
-/*   Updated: 2025/05/18 23:10:39 by dcastor          ###   ########.fr       */
+/*   Updated: 2025/05/22 11:58:42 by dcastor          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stdlib.h>
 
-int	*ft_dup_int_arr(int arr[], size_t size)
+int	*ft_dup_int_arr(int arr[], unsigned int size)
 {
-	int		*new_arr;
-	size_t	i;
+	int				*new_arr;
+	unsigned int	i;
 
 	new_arr = malloc(sizeof(int) * size);
 	if (!new_arr)
